move meta-header of the action sub-module to its folder and update the range meta-header
