+++ include/seqan3/range/action/all.hpp
