+++ include/seqan3/range/all.hpp
@@ -41,7 +41,7 @@
 #pragma once
 
 #include <seqan3/range/concept.hpp>
-#include <seqan3/range/action.hpp>
+#include <seqan3/range/action/all.hpp>
 #include <seqan3/range/container.hpp>
 #include <seqan3/range/view.hpp>
